moved all classes out of main
