+++ lib/core/models/transaction_status.dart
@@ -1,0 +1,5 @@
+enum TransactionStatus {
+  pending,
+  confirmed,
+  failed,
+}
