+++ lib/core/models/transaction_type.dart
@@ -1,0 +1,4 @@
+enum TransactionType {
+  sent,
+  received,
+}
