+++ lib/features/main/screens/wallet_main.dart
@@ -3,11 +3,12 @@ import 'package:flutter/services.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'package:resonance_network_wallet/account_profile.dart';
 import 'package:resonance_network_wallet/core/extensions/color_extensions.dart';
+import 'package:resonance_network_wallet/core/models/transaction_model.dart';
+import 'package:resonance_network_wallet/core/models/transaction_type.dart';
 import 'package:resonance_network_wallet/core/services/substrate_service.dart';
 import 'package:resonance_network_wallet/features/components/transaction_list_item.dart';
 import 'package:resonance_network_wallet/features/main/screens/send_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:resonance_network_wallet/main.dart';
 
 class WalletMain extends StatefulWidget {
   const WalletMain({super.key});
